requirements.txt
asgiref==3.7.2
attrs==23.2.0
catboost==1.1.1
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
Django==3.2.16
django-bootstrap5==22.2
django-cleanup==8.1.0
django-cors-headers==3.13.0
django-filter==21.1
djangorestframework==3.15.1
drf-yasg==1.21.7
flake8==5.0.4
flake8-docstrings==1.7.0
fonttools==4.50.0
graphviz==0.20.3
gunicorn==20.0.4
inflection==0.5.1
iniconfig==2.0.0
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.3
mccabe==0.7.0
numpy==1.26.4
packaging==24.0
pandas==1.4.4
Pillow==9.3.0
plotly==5.20.0
pluggy==1.4.0
psycopg2-binary==2.9.9
py==1.11.0
pycodestyle==2.9.1
pydocstyle==6.3.0
pyflakes==2.5.0
pyparsing==3.1.2
pytest==7.1.3
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
scikit-learn==1.0.2
scipy==1.12.0
six==1.16.0
snowballstemmer==2.2.0
sorl-thumbnail==12.10.0
sqlparse==0.4.4
tenacity==8.2.3
threadpoolctl==3.4.0
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
